message commit

+++ backend/src/main/java/org/desafioestagio/backend/service/ClienteService.java
@@ -1,14 +1,17 @@
 package org.desafioestagio.backend.service;
 
+import org.desafioestagio.backend.exception.ClienteJaCadastradoException;
 import org.desafioestagio.backend.model.Cliente;
-import org.desafioestagio.backend.controller.ClienteRepository;
+import org.desafioestagio.backend.repository.ClienteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.desafioestagio.backend.exception.ClienteNaoEncontradoException;
+
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ClienteService {
+
     @Autowired
     private ClienteRepository clienteRepository;
 
@@ -16,18 +19,22 @@ public class ClienteService {
         return clienteRepository.findAll();
     }
 
-    public Optional<Cliente> buscarPorId(Long id) {
-        return clienteRepository.findById(id);
+    public Cliente buscarPorId(Long id) {
+        return clienteRepository.findById(id)
+                .orElseThrow(() -> new ClienteNaoEncontradoException("Cliente não encontrado com id: " + id));
     }
 
     public Cliente salvar(Cliente cliente) {
         if (clienteRepository.existsByCpfCnpj(cliente.getCpfCnpj())) {
-            throw new RuntimeException("CPF/CNPJ já cadastrado");
+            throw new ClienteJaCadastradoException("CPF/CNPJ já cadastrado");
         }
         return clienteRepository.save(cliente);
     }
 
     public void excluir(Long id) {
+        if (!clienteRepository.existsById(id)) {
+            throw new ClienteNaoEncontradoException("Cliente não encontrado para exclusão com id: " + id);
+        }
         clienteRepository.deleteById(id);
     }
 }
